+++ Erlang-Legacy/Assets/Core/Cinematographics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 273e0964210ae5e438100ea698cd4d00
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
