--- Erlang-Legacy/Assets/StartWithJumpPosse.cs
@@ -1,15 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using Core.Player.Controller;
-
-public class StartWithJumpPosse : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        PlayerController.Instance.Animator.SetBool("jumping",true);
-        PlayerController.Instance.Animator.SetTrigger("jump");
-    }
-
-}
